aj.luo@mail.utoronto.ca umayrah.chonee@mail.utoronto.ca
decoupled the controllers
