+++ src/Main2/Main.java
@@ -1,11 +1,10 @@
-import login_functions.GetIDandPasswords;
-import login_functions.WelcomePageController;
+import login_functions.WelcomePageGUI;
 
 import java.io.IOException;
 
 public class Main {
     public static void main(String[] args) throws IOException, ClassNotFoundException {
-        WelcomePageController welcomePageController = new WelcomePageController();
+        WelcomePageGUI welcomePageGUI = new WelcomePageGUI();
         
 
     }
